+++ package.json
@@ -42,7 +42,7 @@
     "eslint": "^8.30.0",
     "express": "^4.18.2",
     "express-validator": "^6.14.2",
-    "jest-mock-extended": "2.0.4",
+    "jest-mock-extended": "^3.0.1",
     "jsonwebtoken": "^8.5.1",
     "prisma": "^4.5.0",
     "supertest": "^6.3.3"
